+++ src/features/settings/screens/Appearance.tsx
@@ -14,7 +14,7 @@ import ListSubheader from '@mui/material/ListSubheader';
 import Switch from '@mui/material/Switch';
 import Link from '@mui/material/Link';
 import { useColorScheme } from '@mui/material/styles';
-import { useLingui } from '@lingui/react/macro';
+import { Trans, useLingui } from '@lingui/react/macro';
 import { useAppThemeContext } from '@/features/theme/AppThemeContext.tsx';
 import { Select } from '@/base/components/inputs/Select.tsx';
 import { MediaQuery } from '@/base/utils/MediaQuery.tsx';
@@ -37,6 +37,8 @@ import { AppStorage } from '@/lib/storage/AppStorage.ts';
 import { useAppTitle } from '@/features/navigation-bar/hooks/useAppTitle.ts';
 import { MANGA_GRID_WIDTH, SERVER_SETTINGS_METADATA_DEFAULT } from '@/features/settings/Settings.constants.ts';
 import { MUI_THEME_MODE_KEY } from '@/lib/mui/MUI.constants.ts';
+import { getISOLanguage } from '@/lib/ISOLanguageUtil.ts';
+import Typography from '@mui/material/Typography';
 
 export const Appearance = () => {
     const { t } = useLingui();
@@ -123,16 +125,25 @@ export const Appearance = () => {
                     <ListItemText
                         primary={t`Language`}
                         secondary={
-                            <>
-                                <span>{t`Feel free to translate the project on`} </span>
+                            <Trans>
+                                Feel free to translate the project on Weblate (
                                 <Link
                                     href="https://hosted.weblate.org/projects/suwayomi/suwayomi-webui"
                                     target="_blank"
                                     rel="noreferrer"
                                 >
-                                    {t`Weblate`}
+                                    WebUI
+                                </Link>{' '}
+                                and{' '}
+                                <Link
+                                    href="https://hosted.weblate.org/projects/suwayomi/suwayomi-server"
+                                    target="_blank"
+                                    rel="noreferrer"
+                                >
+                                    Server
                                 </Link>
-                            </>
+                                )
+                            </Trans>
                         }
                     />
                     <Select
@@ -145,10 +156,23 @@ export const Appearance = () => {
                                 makeToast(t`Could not load language`, 'error', getErrorMessage(e));
                             });
                         }}
+                        renderValue={(selected) => {
+                            const selectedLanguage =
+                                i18nResources.find((language) => language === selected) ?? selected;
+
+                            return languageCodeToName(selectedLanguage);
+                        }}
                     >
                         {i18nResources.map((language) => (
-                            <MenuItem key={language} value={language}>
+                            <MenuItem
+                                key={language}
+                                value={language}
+                                sx={{ flexDirection: 'column', alignItems: 'flex-start' }}
+                            >
                                 {languageCodeToName(language)}
+                                <Typography color="textSecondary" variant="body2">
+                                    {getISOLanguage(language)?.name}
+                                </Typography>
                             </MenuItem>
                         ))}
                     </Select>